# packages/client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pario API",
    "description": "Ontology-first digital twins runtime API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "parioCsrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-pario-csrf",
        "description": "Required for authenticated mutating requests. Use the csrfToken returned by GET /api/auth/session."
      }
    },
...
  },
...
  "paths": {
...
    "/api/auth/sign-out": {
      "post": {
        "summary": "Sign out current auth session",
        "tags": ["Auth"],
        "operationId": "signOut",
        "security": [
          {
            "parioCsrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": ["success"],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/invitations/{invitationId}/revoke": {
      "post": {
        "summary": "Revoke an auth invitation",
        "tags": ["Auth"],
        "operationId": "revokeAuthInvitation",
        "security": [
          {
            "parioCsrf": []
          }
        ],
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "invitation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "groupIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "status": {
                          "type": "string",
                          "enum": ["pending", "accepted", "revoked"]
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string"
                        },
                        "acceptedAt": {
                          "type": "string"
                        },
                        "revokedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "groupIds",
                        "status",
                        "createdAt",
                        "updatedAt",
                        "expiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["invitation"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Response for status 400",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}